+++ sed/loader/flash/instruments.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from dask import dataframe as dd
+
+
+def wespe_convert(df: dd.DataFrame, df_timed: dd.DataFrame) -> tuple[dd.DataFrame, dd.DataFrame]:
+    df
+    df_timed
+    raise NotImplementedError("This function is not implemented yet.")
